backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
supabase==2.3.4

# Async Tasks
celery==5.3.6
redis==5.0.1

# AI/ML Services
openai==1.10.0
audiocraft==1.2.0
torch==2.1.2
torchaudio==2.1.2
transformers==4.37.0
matchering==2.0.8
librosa==0.10.1
soundfile==0.12.1

# Audio Processing
pydub==0.25.1
scipy==1.12.0
numpy==1.26.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Monitoring & Logging
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pillow==10.2.0
